ci: Break up the consumer chain interchaintests into user stories
